Don't use -Werror on Darwin
As of this writing (2018-10-20) there seems to be an issue with the
latest version of Xcode's command line tools that cause warnings of the
format

ld: warning: text-based stub file ... and library file ... are out of
sync. Falling back to library file for linking.

This causes problems here because mkmf uses -Werror to verify the
validity of an -L location, then fails on the above warnings, and
disregards even valid locations supplied by `R CMD config --ldflags`.

As a temporary fix, force no -Werror on Darwin.