Fix typo PROTECr => PROTECT
I think this was accidentally introduced in
39c788799d3fd53c1266206490fbdd56743a6918. I believe it hasn't caused
problems because if we're in this part of the code we are raising a Ruby
exception anyway (`rb_exc_raise`).

If you try to build this with

  -Werror,-Wimplicit-function-declaration

then you get an error

  error: implicit declaration of function 'PROTECr' is invalid in C99

so apparently it's been treating this as an implicit declaration all
this time. This came up because attempts to install this gem with more
recent versions of gem failed; I suspect somewhere along the way ruby
(in mkmf, presumably?) has started building Makefiles with
`-Wimplicit-function-declaration`, which all in all is a good thing.

Other changes beside this one are just automatic whitespace cleanup.